+++ real/src/App.js
@@ -1,6 +1,6 @@
 import logo from './logo.svg';
 import './App.css';
-import Mission from './mission';
+import Mission from './Missionandvission';
 
 function App() {
   return (
