+++ real/src/Missionandvission/index.js
@@ -1,0 +1,32 @@
+import './index.css'
+const Mission = () =>{
+    return(
+        <div className='Missionandvision'>
+            <div className='mission'>
+        <h3> Mission</h3>
+        <p>Real Fashion is an online B2B platform
+           that offers second hand traders with real
+           time analytics on the market trends in the 
+           most simple and efficient way. We offer a
+           personalized customer service. Our mission
+           is to make a difference through our branding
+           to stay ahead of fashion trends and provide
+           the best services to our customers.
+        </p>
+        </div>
+        <div className='vision'>
+            <h3>Vision</h3>
+            <p>By using the latest technology, we aim to
+               facilitate real time analysis on the market
+               trends to our users. Real Fashion aim to be a 
+               pioneer for the fashion industry and help in
+               facilitating the trade of the second hand 
+               clothes in Kenya. We strive to be a caring and 
+               well-managed organisation  for our customers.
+            </p>
+        </div>
+        </div>
+
+    )
+}
+export default Mission;
