--- real/src/mission/index.css
@@ -1,47 +1,0 @@
-.Missionandvision{
-    display: flex;
-    gap: 15%;
-    height: 980px;
-    width: 1850px;
-    background-repeat: no-repeat;
-    background-size: cover;
-}
-.mission{
-    background-color: pink;
-    height: 260px;
-    width: 600px;
-    margin-top: 19%;
-    margin-left: 10%;
-    font-size: larger;
-}
-.vision{
-    background-color: rgb(255, 192, 203);
-    height: 260px;
-    width: 600px;
-    margin-top: 19%;
-    font-size: larger;
-}
-
-@media screen and (max-width: 650px){
-    .Missionandvision{
-        display: grid;
-        width: 280px;
-        /* height: 900px; */
-        grid-template-columns: 1 1fr;
-        background-size: cover;
-    }
-    .mission{
-        background-color: pink;
-        height: 260px;
-        width: 200px;
-        font-size: small;
-        margin-left: 14%;
-    }
-    .vision{
-        background-color: pink;
-        height: 260px;
-        width: 200px;
-        font-size: small;
-        margin-left: 14%;
-    }
-}
