--- real/src/mission/index.js
@@ -1,34 +1,0 @@
-import './index.css'
-const Mission = () =>{
-    return(
-        <div className='Missionandvision'>
-            <div className='mission'>
-        <h3> Mission</h3>
-        <p>Real Fashion is an online B2B platform that offers second hand
-            <br></br>
-            <br></br>
-         traders with real rime analytics on the market trends in the most
-            <br></br>
-            <br></br>
-            simple and efficient way. We offer a personalized customer service.
-        </p>
-        </div>
-        <div className='vision'>
-            <h3>Vision</h3>
-            <p>By using the latest technology, we aim to facilitate real time
-                <br></br>
-                <br></br>
-                analysis on the market trends to our users. Real Fashion will
-                <br></br>
-                <br></br>
-                be a pioneer for the fashion industry and facilitating the
-                <br></br>
-                <br></br>
-                trade of the second hand clothes in Kenya.
-            </p>
-        </div>
-        </div>
-
-    )
-}
-export default Mission;
